+++ CHANGES.md
@@ -12,13 +12,15 @@ NEXT MILESTONE
 - Enable hot variable addition to propagator
 - Move nogood recording from solution and restart from constraint pkg to monitor +(revamp) + SMF API (#261)
 - Explanations are enabled on initial propagation (#247)
-- Improvment of the solution object (#254)
+- Improvement of the solution object (#254)
 - add DBT for explanations
 - Improve explanation engine
 - Add an additional "worldPush" instruction after initial propagation to be coherent with restarts (#55)
 - Remove restarts from ABS (#282)
 - one-shot decision are not reinitialized (#283)
 - Remove Propagation count (#284)
+- Change the default search strategy (#290)
+- Add possibility to complete the declared search strategy (#291)
 
 Bug fixes: #168, #221, #239, #262, #264, #265, #269, #273, #275, #281